Fixed bar title not displaying correct information
There was a problem when hovering the mouse over a chart bar. Instead of
showing the details for that specific item/day, it showed the values for
all the items on that day. This was caused by the fact that tabcount var
was not incremented properly. I've moved the incrementation of this
variable inside the loadCharts function as this was the only place where
this was needed.